+++ packages/web/src/javascripts/Components/ApplicationView/ApplicationView.tsx
@@ -27,6 +27,7 @@ import ApplicationProvider from '../ApplicationProvider'
 import CommandProvider from '../CommandProvider'
 import PanesSystemComponent from '../Panes/PanesSystemComponent'
 import DotOrgNotice from './DotOrgNotice'
+import LinkingControllerProvider from '@/Controllers/LinkingControllerProvider'
 
 type Props = {
   application: WebApplication
@@ -177,62 +178,61 @@ const ApplicationView: FunctionComponent<Props> = ({ application, mainApplicatio
               application={application}
               featuresController={viewControllerManager.featuresController}
             >
-              <div className={platformString + ' main-ui-view sn-component h-full'}>
-                <FileDragNDropProvider
-                  application={application}
-                  featuresController={viewControllerManager.featuresController}
-                  filesController={viewControllerManager.filesController}
-                >
-                  <PanesSystemComponent />
-                </FileDragNDropProvider>
-
-                <>
-                  <Footer application={application} applicationGroup={mainApplicationGroup} />
-                  <SessionsModal application={application} viewControllerManager={viewControllerManager} />
-                  <PreferencesViewWrapper viewControllerManager={viewControllerManager} application={application} />
-                  <RevisionHistoryModal
+              <LinkingControllerProvider controller={viewControllerManager.linkingController}>
+                <div className={platformString + ' main-ui-view sn-component h-full'}>
+                  <FileDragNDropProvider
                     application={application}
-                    historyModalController={viewControllerManager.historyModalController}
-                    notesController={viewControllerManager.notesController}
-                    selectionController={viewControllerManager.selectionController}
-                    subscriptionController={viewControllerManager.subscriptionController}
-                  />
-                </>
-
-                {renderChallenges()}
-
-                <>
-                  <NotesContextMenu
-                    application={application}
-                    navigationController={viewControllerManager.navigationController}
-                    notesController={viewControllerManager.notesController}
-                    linkingController={viewControllerManager.linkingController}
-                    historyModalController={viewControllerManager.historyModalController}
-                  />
-                  <TagContextMenuWrapper
-                    navigationController={viewControllerManager.navigationController}
                     featuresController={viewControllerManager.featuresController}
-                  />
-                  <FileContextMenuWrapper
                     filesController={viewControllerManager.filesController}
-                    selectionController={viewControllerManager.selectionController}
-                  />
-                  <PurchaseFlowWrapper application={application} viewControllerManager={viewControllerManager} />
-                  <ConfirmSignoutContainer
-                    applicationGroup={mainApplicationGroup}
-                    viewControllerManager={viewControllerManager}
-                    application={application}
-                  />
-                  <ToastContainer />
-                  <FilePreviewModalWrapper application={application} viewControllerManager={viewControllerManager} />
-                  <PermissionsModalWrapper application={application} />
-                  <ConfirmDeleteAccountContainer
-                    application={application}
-                    viewControllerManager={viewControllerManager}
-                  />
-                </>
-                {application.routeService.isDotOrg && <DotOrgNotice />}
-              </div>
+                  >
+                    <PanesSystemComponent />
+                  </FileDragNDropProvider>
+                  <>
+                    <Footer application={application} applicationGroup={mainApplicationGroup} />
+                    <SessionsModal application={application} viewControllerManager={viewControllerManager} />
+                    <PreferencesViewWrapper viewControllerManager={viewControllerManager} application={application} />
+                    <RevisionHistoryModal
+                      application={application}
+                      historyModalController={viewControllerManager.historyModalController}
+                      notesController={viewControllerManager.notesController}
+                      selectionController={viewControllerManager.selectionController}
+                      subscriptionController={viewControllerManager.subscriptionController}
+                    />
+                  </>
+                  {renderChallenges()}
+                  <>
+                    <NotesContextMenu
+                      application={application}
+                      navigationController={viewControllerManager.navigationController}
+                      notesController={viewControllerManager.notesController}
+                      linkingController={viewControllerManager.linkingController}
+                      historyModalController={viewControllerManager.historyModalController}
+                    />
+                    <TagContextMenuWrapper
+                      navigationController={viewControllerManager.navigationController}
+                      featuresController={viewControllerManager.featuresController}
+                    />
+                    <FileContextMenuWrapper
+                      filesController={viewControllerManager.filesController}
+                      selectionController={viewControllerManager.selectionController}
+                    />
+                    <PurchaseFlowWrapper application={application} viewControllerManager={viewControllerManager} />
+                    <ConfirmSignoutContainer
+                      applicationGroup={mainApplicationGroup}
+                      viewControllerManager={viewControllerManager}
+                      application={application}
+                    />
+                    <ToastContainer />
+                    <FilePreviewModalWrapper application={application} viewControllerManager={viewControllerManager} />
+                    <PermissionsModalWrapper application={application} />
+                    <ConfirmDeleteAccountContainer
+                      application={application}
+                      viewControllerManager={viewControllerManager}
+                    />
+                  </>
+                  {application.routeService.isDotOrg && <DotOrgNotice />}
+                </div>
+              </LinkingControllerProvider>
             </PremiumModalProvider>
           </ResponsivePaneProvider>
         </AndroidBackHandlerProvider>